+++ app-backend/backsplash-server/src/main/scala/com/rasterfoundry/backsplash/Main.scala
@@ -2,6 +2,7 @@ package com.rasterfoundry.backsplash.server
 
 import com.rasterfoundry.database.{
   LayerAttributeDao,
+  SceneDao,
   SceneToProjectDao,
   ToolRunDao
 }
@@ -23,7 +24,7 @@ import org.http4s._
 import org.http4s.circe._
 import org.http4s.dsl.io._
 import org.http4s.headers._
-import org.http4s.server.middleware.{AutoSlash, CORS, CORSConfig, GZip, Timeout}
+import org.http4s.server.middleware.{AutoSlash, CORS, CORSConfig, Timeout}
 import org.http4s.server.blaze.BlazeServerBuilder
 import org.http4s.server.Router
 import org.http4s.syntax.kleisli._
@@ -126,10 +127,22 @@ object Main extends IOApp with HistogramStoreImplicits with LazyLogging {
                             toolStoreImplicits,
                             xa).routes))
 
+  val sceneMosaicService: HttpRoutes[IO] =
+    authenticators.tokensAuthMiddleware(
+      AuthedAutoSlash(
+        new SceneService(SceneDao(),
+                         mtr,
+                         mosaicImplicits,
+                         LayerAttributeDao(),
+                         xa).routes
+      )
+    )
+
   val httpApp =
     Router(
-      "/" -> mtr.middleware(GZip(withCORS(withTimeout(mosaicService)))),
-      "/tools" -> mtr.middleware(GZip(withCORS(withTimeout(analysisService)))),
+      "/" -> mtr.middleware(withCORS(withTimeout(mosaicService))),
+      "/scenes" -> mtr.middleware(withCORS(withTimeout(sceneMosaicService))),
+      "/tools" -> mtr.middleware(withCORS(withTimeout(analysisService))),
       "/healthcheck" -> AutoSlash(new HealthcheckService[IO]().routes)
     )